requirements.txt
flask == 2.3.3
yfinance >= 0.2.66
numpy
pandas
tensorflow-cpu
scikit-learn
